transfered font printing to external file
