+++ main.go
@@ -12,17 +12,9 @@ import (
 	"edchja.de/tic-tac-toe/fonts"
 )
 
-var gameBoardString = `
-|   |   |   |
-–––––––––––––
-|   |   |   |
-–––––––––––––
-|   |   |   |`
-
 const (
-	refreshRate = 100 * time.Millisecond
-	player1     = 1
-	player2     = 2
+	player1 = 1
+	player2 = 2
 )
 
 var gameBoard = make([][]int, 3)
@@ -33,33 +25,15 @@ var drawValues = []string{" ", "X", "O"}
 var reader = bufio.NewReader(os.Stdin)
 var xWon, oWon, tie bool
 
-// TODO Computerlogik implementieren (wenn 2 X in einer reihe ist, dazwischen O setzen)
-// TODO	Cursor resetting einfügen
+// TODO:	Cursor resetting einfügen
+// TODO:	Refactor to structs
+// TODO:	Refactor to DRY
+// TODO:	AI never loses
 
 func main() {
 	var gameState bool
 
-	// propagates array with 0.
-	for row := range gameBoard {
-		gameBoard[row] = make([]int, 3)
-		for col := range gameBoard[row] {
-			gameBoard[row][col] = 0
-		}
-	}
-
-	// row := 0
-	// col := 0
-	// gameBoardArr[0] = make([]rune, 13)
-	// for _, r := range gameBoardString {
-	// 	if r == '\n' {
-	// 		row++
-	// 		gameBoardArr[row] = make([]rune, 13)
-	// 		col = 0
-	// 		continue
-	// 	}
-	// 	gameBoardArr[row][col] = r
-	// 	col++
-	// }
+	fillArray()
 
 	printBoard()
 
@@ -69,21 +43,15 @@ func main() {
 		processInput(input, player1)
 
 		computerInput := getRandomNumber(1, 9)
-		// fmt.Println("{computer:", computerInput, "}")
 		processInput(computerInput, player2)
 
 		gameState = gamelogic()
-		// fmt.Println(gameState)
 
 		printBoard()
 		printWinningScreen()
 	}
 }
 
-// ? Refactor to DRY
-// TODO: Refactor to structs
-// TODO: AI never loses
-// Docker
 func gamelogic() bool {
 	var (
 		// Check all rows.
@@ -130,7 +98,6 @@ func gamelogic() bool {
 		tie = true
 		return tie
 	}
-
 	return false
 }
 
@@ -164,6 +131,16 @@ func processInput(input int, player int) {
 	} else if player == 2 {
 		computer := getRandomNumber(1, 9)
 		processInput(computer, player2)
+	}
+}
+
+func fillArray() {
+	// propagates array with 0.
+	for row := range gameBoard {
+		gameBoard[row] = make([]int, 3)
+		for col := range gameBoard[row] {
+			gameBoard[row][col] = 0
+		}
 	}
 }
 
@@ -211,62 +188,11 @@ func printBoard() {
 func printWinningScreen() {
 	switch {
 	case xWon:
-		printFont(fontArr, fonts.XWonFont, 45)
+		fonts.PrintFont(fontArr, fonts.XWonFont, 45)
 	case oWon:
-		printFont(fontArr, fonts.OWonFont, 45)
+		fonts.PrintFont(fontArr, fonts.OWonFont, 45)
 	case tie:
-		printFont(fontArr, fonts.TieFont, 25)
-	}
-}
-
-func printFont(arr [][]rune, font string, size int) {
-	row := 0
-	col := 0
-	arr[0] = make([]rune, size)
-	for _, r := range font {
-		if r == '\n' {
-			row++
-			arr[row] = make([]rune, size)
-			col = 0
-			continue
-		}
-		arr[row][col] = r
-		col++
-	}
-
-	go func() {
-		first := true
-		for {
-			if !first {
-				for i := 0; i < len(arr); i++ {
-					up()
-					clearLine()
-				}
-			}
-			for _, line := range arr {
-				fmt.Println(string(line))
-			}
-			first = false
-			time.Sleep(refreshRate)
-		}
-	}()
-
-	go flow(arr)
-
-	for {
-	}
-}
-
-func flow(arr [][]rune) {
-	for {
-		for row := range arr {
-			tempLastIndex := arr[row][len(arr[row])-1]
-			for col := len(arr[row]) - 1; col > 0; col-- {
-				arr[row][col] = arr[row][col-1]
-			}
-			arr[row][0] = tempLastIndex
-		}
-		time.Sleep(100 * time.Millisecond)
+		fonts.PrintFont(fontArr, fonts.TieFont, 25)
 	}
 }
 
@@ -279,19 +205,4 @@ func getRandomNumber(min, max int) int {
 	} else {
 		return getRandomNumber(min, max)
 	}
-}
-
-// Bewegt den Cursor eine Zeile nach oben.
-func up() {
-	fmt.Print("\033[A")
-}
-
-// Bewegt den Cursor eine Zeile runter.
-func down() {
-	fmt.Print("\033[B")
-}
-
-// Löscht die aktuelle Zeile.
-func clearLine() {
-	fmt.Print("\033[G\033[K")
 }
